pcd: inline the grind into fuse
